Move cron endpoint to ggrc module
